Pushing the delete stuff I've done

+++ vue/src/components/ListOfLists.vue
@@ -29,12 +29,17 @@ export default {
   },
   methods: {
     deleteList(listId) {
+      
+      if (confirm("Are you sure you want to delete this list?")) {
       listService.deleteList(listId)
         .then( response => {
           if (response.status == 204) {
             this.$store.commit("DELETE_LIST", listId);
           }
-        })
+        }).catch((err) => {
+          console.error(err + " problem deleting list!");
+        });
+      }
     }
   }
 };
